+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 v0.6.4 (TBD)
 --------------------
  - [Machine Learning] Mean-square error prints in scientific notation if not between 1 and 1000
+ - [DICOM Import] Ensure RxParser.fx_count is an integer (GH Issue 18)
 
 
 v0.6.3 (2019.12.15)